fix(coin): sum().longValue()로 QueryDSL Integer→Long 타입 변환 명시

+++ SClass-Domain/src/main/kotlin/com/sclass/domain/domains/coin/repository/CoinTransactionCustomRepositoryImpl.kt
@@ -34,7 +34,7 @@ class CoinTransactionCustomRepositoryImpl(
                     Projections.constructor(
                         CoinTransactionGroupDto::class.java,
                         coinTransaction.type,
-                        coinTransaction.amount.sum(),
+                        coinTransaction.amount.sum().longValue(),
                         coinTransaction.referenceId,
                         coinTransaction.description.max(),
                         coinTransaction.createdAt.min(),
